+++ zk_shell/tests/test_basic_cmds.py
@@ -368,3 +368,15 @@ class BasicCmdsTestCase(ShellTestCase):
         # now remove it
         self.shell.onecmd("reconfig remove 100")
         self.assertNotIn(joining, self.output.getvalue())
+
+    def test_time(self):
+        self.shell.onecmd("time 'ls /'")
+        self.assertIn("Took", self.output.getvalue())
+        self.assertIn("seconds", self.output.getvalue())
+
+    def test_create_async(self):
+        self.shell.onecmd(
+            "create %s/foo bar ephemeral=false sequence=false recursive=false async=true" % (
+                self.tests_path))
+        self.shell.onecmd("exists %s/foo" % self.tests_path)
+        self.assertIn("numChildren=0", self.output.getvalue())