+++ CHANGES.md
@@ -6,7 +6,7 @@ ChangeLog
 
 Bug Handling
 ************
--
+- fix doc error in `sleep`'s documentation
 
 Features
 ********
@@ -17,6 +17,8 @@ Features
   ```
   (CONNECTED) /> ls / watch=true
   ```
+- new command `time`: allows time the execution of the given commands
+- the `create` command now supports async mode (async=true)
 
 1.1.0 (2015-06-17)
 --------------------